LayoutTests/platform/mac/accessibility/slider-allows-title-ui-element.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<link rel="stylesheet" href="../../../fast/js/resources/js-test-style.css">
<script>
var successfullyParsed = false;
</script>
<script src="../../../fast/js/resources/js-test-pre.js"></script>
</head>
<body id="body">

<label for="slider" tabindex="0" id="title">Color:</label>
<input id="slider" type="range" min="0" value="0" max="6">

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that sliders are allowed to have title UI elements.");

    if (window.accessibilityController) {

        document.getElementById("slider").focus();
        var slider = accessibilityController.focusedElement;

        document.getElementById("title").focus();
        var title = accessibilityController.rootElement.childAtIndex(0).childAtIndex(0).childAtIndex(0);

        shouldBeTrue("slider.titleUIElement().isEqual(title)");
        shouldBe("title.role", "'AXRole: AXGroup'");
        shouldBe("title.childAtIndex(0).role", "'AXRole: AXStaticText'");
    }

    successfullyParsed = true;
</script>

<script src="../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
